# Restore:  improve thread pool starvation fix by disabling asynchronous I/O