--- a/src/NuGet.Core/NuGet.Protocol/HttpSource/HttpCacheUtility.cs
+++ b/src/NuGet.Core/NuGet.Protocol/HttpSource/HttpCacheUtility.cs
@@ -3,7 +3,6 @@
 
 using System;
 using System.IO;
-using System.Linq;
 using System.Net.Http;
 using System.Threading;
 using System.Threading.Tasks;
@@ -67,44 +66,24 @@ namespace NuGet.Protocol
             // The update of a cached file is divided into two steps:
             // 1) Delete the old file.
             // 2) Create a new file with the same name.
-
-            // Some FileStream operations on Windows are synchronous even though it may not seem so.
-            // The HTTP stack rewrite in .NET Core 2.1 introduced circumstances whereby these
-            // synchronous FileStream calls will keep an IO completion thread busy and block other
-            // HTTP requests from completing.  The immediate solution is to perform write and read
-            // operations on separate streams, but only on .NET Core where the problem exists.
-            // See https://github.com/dotnet/corefx/issues/31914 for details.
-            const int writeBufferSize =
-#if IS_CORECLR
-                1;  // This disables write buffering.
-#else
-                BufferSize;
-#endif
-
             using (var fileStream = new FileStream(
                 result.NewFile,
                 FileMode.Create,
-                FileAccess.Write,
+                FileAccess.ReadWrite,
                 FileShare.None,
-                writeBufferSize,
-                useAsync: true))
+                BufferSize))
             {
                 using (var networkStream = await response.Content.ReadAsStreamAsync())
                 {
                     await networkStream.CopyToAsync(fileStream, BufferSize, cancellationToken);
                 }
-            }
 
-            using (var fileStream = new FileStream(
-                result.NewFile,
-                FileMode.Open,
-                FileAccess.Read,
-                FileShare.None,
-                BufferSize,
-                useAsync: true))
-            {
                 // Validate the content before putting it into the cache.
-                ensureValidContents?.Invoke(fileStream);
+                if (ensureValidContents != null)
+                {
+                    fileStream.Seek(0, SeekOrigin.Begin);
+                    ensureValidContents.Invoke(fileStream);
+                }
             }
 
             if (File.Exists(result.CacheFile))
@@ -141,8 +120,7 @@ namespace NuGet.Protocol
                 FileMode.Open,
                 FileAccess.Read,
                 FileShare.Read | FileShare.Delete,
-                BufferSize,
-                useAsync: true);
+                BufferSize);
         }
     }
 }