Resolve https://github.com/NuGet/Home/issues/7314.

The fix is to disable asynchronous I/O in the `FileSystem` class.

Ignoring the [previous change](https://github.com/NuGet/NuGet.Client/pull/2430), this change improves restore performance under .NET Core 2.1+ by ~70% using .NET Core 2.2 Preview 3 (2.2.100-preview3-009430) and the repro for https://github.com/dotnet/corefx/issues/31914.

Trial | Baseline (seconds) | With `useAsync == false` (seconds)
-- | -- | --
1 | 293.6 | 66.99
2 | 280.06 | 65.87
3 | 276.75 | 73.42
4 | 273.04 | 81.44
5 | 287.19 | 102.84
6 | 269.03 | 72.28
7 | 288.72 | 100.19
8 | 298.51 | 109.87
9 | 275.34 | 78.8
10 | 268.6 | 84.41
Average | 281.084 | 83.611

All baseline trials ultimately failed with:

    The SSL connection could not be established, see inner exception.
    Authentication failed because the remote party has closed the transport stream.

None of the trials with the fix failed.

This fix was intended for .NET Core 2.1+ as its HTTP rewrite was the catalyst for this performance regression.  However, I tested the same fix with NuGet.exe and found a ~44% improvement in restore performance.

Trial | Baseline (seconds) | With `useAsync == false` (seconds)
-- | -- | --
1 | 85.8 | 47.2
2 | 78.0 | 42.2
3 | 84.0 | 45.5
4 | 88.2 | 45.5
5 | 96.0 | 49.6
6 | 89.4 | 66.0
7 | 96.0 | 45.4
8 | 81.6 | 47.1
9 | 84.0 | 45.6
10 | 82.2 | 47.7
Average | 86.5 | 48.2