@loic-sharma 2018-10-22
> However, I tested the same fix with NuGet.exe and found a ~44% improvement in restore performance.

Do we know why removing async improves performance? Is this also due to the HTTP rewrite you mentioned?